feat: Create ModerationAction and Appeal DTOs (T186 & T187)
Implement proper Data Transfer Objects for moderation actions and appeals:

- Create moderation-action.dto.ts with DTOs for creating/managing actions:
  * CreateActionRequest, ApproveActionRequest, RejectActionRequest
  * ModerationActionResponse, ModerationActionDetailResponse
  * ListActionsResponse, CoolingOffPromptRequest/Response
  * ModeratorInfo type for approval tracking

- Create appeal.dto.ts with DTOs for managing appeals:
  * CreateAppealRequest, ReviewAppealRequest
  * AppealResponse, PendingAppealResponse, ListAppealResponse
  * ReviewerInfo type for appeal review tracking

- Refactor ModerationActionsService to use DTOs:
  * Import types from new DTO files
  * Re-export DTOs for backward compatibility
  * Update method signatures to use typed responses

- Update ModerationController to import DTOs:
  * Use DTO types in endpoint signatures
  * Maintain backward compatibility

- Create dto/index.ts for centralized DTO exports

Changes support the user warning system (actionType: 'warn') which is a
NON_PUNITIVE severity action that notifies users of policy violations.

Fixes #182
