+++ services/moderation-service/src/dto/index.ts
@@ -1,0 +1,25 @@
+/**
+ * Moderation Service DTOs
+ * Centralized exports for all Data Transfer Objects
+ */
+
+export type {
+  CreateActionRequest,
+  ApproveActionRequest,
+  RejectActionRequest,
+  ModeratorInfo,
+  ModerationActionResponse,
+  ModerationActionDetailResponse,
+  ListActionsResponse,
+  CoolingOffPromptRequest,
+  CoolingOffPromptResponse,
+} from './moderation-action.dto.js';
+
+export type {
+  CreateAppealRequest,
+  ReviewAppealRequest,
+  ReviewerInfo,
+  AppealResponse,
+  PendingAppealResponse,
+  ListAppealResponse,
+} from './appeal.dto.js';
